Move the generateLicenseXml, createIssue, parseXmlString and get_xml methods to utils file from app views
